Migrating from MongoDB to TokuMX
TokuMX ( http://www.tokutek.com/products/tokumx-for-mongodb/ ) is a replacement for MongoDB that changes all the storage code for better performance, concurrency, and compression. Because the data format is different than in MongoDB, migration requires a full dump-and-reload, and because the replication format is different, it can be hard to migrate a live production system.

In TokuMX 1.0.3, we are introducing a tool to make this process easier. It replays a vanilla MongoDB replication oplog on TokuMX. This makes the migration process much simpler, more seamless, and it can be done with next to no downtime.

This asciicast is an executable presentation. You can see the org file I walk through and my python script "application" at leifwalsh/tokumx-migration-demo ( http://github.com/leifwalsh/tokumx-migration-demo/ ) . You can clone this repo, read the presentation, and follow along as you walk through the process yourself, in a terminal or using Babel ( http://orgmode.org/worg/org-contrib/babel/ ) .